This commit improves the parsing of HTTP trigger headers that could be in JSON format.
Instead of relying on try-catch blocks which attempt to deserialize the header blindly, this diff first checks if the header string starts and ends with a JSON-compatible syntax (curly braces) before attempting to deserialize. In the event that parsing fails, the header is treated as a comma-separated list of simplified event keys.