Update HtmxResponseHeaders with new constants and methods
The HtmxResponseHeaders class has been updated to include new constants and a new method 'Reselect'. The constants have been rearranged for better reference based on the htmx documentation. The 'Reselect' method allows to choose specific parts of the response for swapping in, overriding an existing 'hx-select' on the triggering element.

#40